Change nodeid to an int.

+++ src/MainNFSD/nfs_init.c
@@ -82,7 +82,7 @@ verifier4 NFS4_write_verifier;	/* NFS V4 write verifier */
 writeverf3 NFS3_write_verifier;	/* NFS V3 write verifier */
 
 /* node ID used to identify an individual node in a cluster */
-ushort g_nodeid = 0;
+int g_nodeid = 0;
 
 nfs_start_info_t nfs_start_info;
